+++ lib/nulldevice.cpp
@@ -2,7 +2,7 @@
 // nulldevice.cpp
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2014-2017  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2014-2018  R. Stange <rsta2@o2online.de>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -29,12 +29,12 @@ CNullDevice::~CNullDevice (void)
 {
 }
 
-int CNullDevice::Read (void *pBuffer, unsigned nCount)
+int CNullDevice::Read (void *pBuffer, size_t nCount)
 {
 	return 0;
 }
 
-int CNullDevice::Write (const void *pBuffer, unsigned nCount)
+int CNullDevice::Write (const void *pBuffer, size_t nCount)
 {
 	return (int) nCount;
 }
