+++ lib/soundbasedevice.cpp
@@ -2,7 +2,7 @@
 // soundbasedevice.cpp
 //
 // Circle - A C++ bare metal environment for Raspberry Pi
-// Copyright (C) 2017  R. Stange <rsta2@o2online.de>
+// Copyright (C) 2017-2018  R. Stange <rsta2@o2online.de>
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -134,7 +134,7 @@ void CSoundBaseDevice::SetWriteFormat (TSoundFormat Format, unsigned nChannels)
 	m_nWriteFrameSize = m_nWriteChannels * m_nWriteSampleSize;
 }
 
-int CSoundBaseDevice::Write (const void *pBuffer, unsigned nCount)
+int CSoundBaseDevice::Write (const void *pBuffer, size_t nCount)
 {
 	assert (m_WriteFormat < SoundFormatUnknown);
